REVIEWME: Replace kvm virtualization by qemu
The actual test is done in a VM in a host which does not accept
other kind of virtualization. In the future this patch should
be reverted or not applied on rebase if the virtualization
required is KVM or modified if the virtualization is any other.

Qemu is only configured for test purposes.