Add xstatic stratum which is dependecy for horizon
Xstatic is just a wrapper to install some packages from an
static build using pip install python.
This is against baserock foundations, so this stratum is just for
testing that these are the packages required.
After that we should try to add the correct packages and maybe
patch horizon to not require the xstatic-* packages.