Neutron: update configuration files to Kilo
This commit updates the Neutron's configuration files to be the factory
versions for the Kilo release.  Our custom configuration will be
re-added in a following commit.

Also install configuration files which will are not going to be modified
in the following commit, in the post-install-commands for the chunk; as
opposite to having them laying around in the repo and installing them
with the install-files configuration extension.